+++ metapp/metapp.ml
@@ -656,6 +656,26 @@ module Types = struct
            [%p? Mty_alias (_, p)]] -> Some p
       | _ -> None
   end
+
+  [%%meta if Sys.ocaml_version >= "4.13.0" then [%stri
+    type variant_representation = Types.variant_representation =
+      | Variant_regular
+      | Variant_unboxed]
+  else [%stri
+    type variant_representation =
+      | Variant_regular
+      | Variant_unboxed]]
+
+  let destruct_type_variant (type_kind : ('lbl, 'cstr) Types.type_kind) :
+    ('cstr list * variant_representation) option =
+    [%meta if Sys.ocaml_version >= "4.13.0" then [%e
+      match type_kind with
+      | Type_variant (list, repr) -> Some (list, repr)
+      | _ -> None]
+    else [%e
+      match type_kind with
+      | Type_variant list -> Some (list, Variant_unboxed)
+      | _ -> None]]
 end
 
 (** {1 With constraint} *)